配置调整
